tests/slack_app/requirements.txt
slack-bolt
python-dotenv
certifi